+++ OpenRA.Mods.D2k/Render/WithCrumbleOverlay.cs
@@ -9,9 +9,11 @@
 #endregion
 
 using OpenRA.Graphics;
+using OpenRA.Mods.RA;
+using OpenRA.Mods.RA.Render;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Render
+namespace OpenRA.Mods.D2k.Render
 {
 	[Desc("Rendered together with the \"make\" animation.")]
 	public class WithCrumbleOverlayInfo : ITraitInfo, Requires<RenderSpritesInfo>
