+++ OpenRA.Mods.D2k/SpriteLoaders/R8Loader.cs
