+++ OpenRA.Mods.D2k/World/D2kResourceLayer.cs
