+++ OpenRA.Mods.D2k/World/PaletteFromScaledPalette.cs
